requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
future==1.0.0
ply==3.11
pyzk==0.9
setuptools==75.9.1
six==1.17.0
sqlparse==0.5.3
thriftpy2==0.5.3
tzdata==2025.2
